Fixed and cleaned up util.Graph
Fixed bug in util.Graph.shortestPath where it was failing to find a path
where it should.  This was achieved by initializing the
shortestPathLength variable to Double.POSITIVE_INFINITY instead of 0d.

Documentation has been updated for the search methods indicating the
output for the condition when no path exists in the graph between the
specified end-points.  The "Exploring..." output lines were removed.
Also, pathLengths no longer have a unit.